requirements.txt
# Web Framework and Server
fastapi
uvicorn[standard]

# Database
sqlalchemy

# AI and LLM
langchain
sentence-transformers
# langchain-openai  # Or 
langchain-google-genai

# Utilities
python-dotenv
pandas
numpy